Switch to libosmium.
All changes necessary for switching from the old Osmium to the new libosmium
(http://osmcode.org/libosmium). This is a more or less one-to-one change
without changing much of the functionality. The only real change is that the
progress display is not supported any more. Libosmium doesn't have support
for that like the old Osmium did and it was slowing down the program quite
a bit, too. Compiling now needs CMake, the old Makefile isn't used any more.
And the code needs a lot more cleanup...